# toteco_openapi.yaml
openapi: 3.0.2
info:
  title: TOTECO API
  description: API to manage publications about products (coffee, tea or toast) like a social media. The idea is to create a community that publishes photos of their toasts with coffee or tea adding the price and being able to rate the products and the establishments.
  version: "1.0"
  contact:
    email: a25540@svalero.com
servers:
  - url: http://localhost:8081
...
paths:
...
  /publications:
    get:
      tags:
        - Publication
      summary: Get all publications
      description: Get all publications
      operationId: getAllPublications
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
                
    post:
      security:
        - bearerAuth: []
      tags:
        - Publication
      summary: Save publication
      description: Save one publication
      operationId: savePublication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicationDTO'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
                    
    delete:
      security:
        - bearerAuth: []
      tags:
        - Publication
      summary: Delete publications
      description: Delete all publications
      operationId: deleteAllPublications
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: string
                example: All publications deleted.
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /publications/{id}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Publication
      summary: Get publication by id
      description: Get one publication by the indicated id
      operationId: getPublicationById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
    
    put:
      security:
        - bearerAuth: []
      tags:
        - Publication
      summary: Update publication
      description: Update an existing publication
      operationId: updatePublication
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicationDTO'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publication'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
                    
    delete:
      security:
        - bearerAuth: []
      tags:
        - Publication
      summary: Delete publication by id
      description: Delete an existing publication
      operationId: deletePublication
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: string
                example: Publication deleted.
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
...
  /publications/date/{date}:
    get:
      tags:
        - Publication
      summary: Get publication by date
      description: Get publications created at the indicated date. The date format is timestamp in miliseconds
      operationId: getPublicationByDate
      parameters:
        - $ref: '#/components/parameters/date'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /publications/date/between:
    get:
      tags:
        - Publication
      summary: Get publication by date between
      description: Get publications created at the indicated dates. The dates format is timestamp in miliseconds
      operationId: getPublicationByDateBetween
      parameters:
        - $ref: '#/components/parameters/minDate'
        - $ref: '#/components/parameters/maxDate'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /publications/price/{price}:
    get:
      tags:
        - Publication
      summary: Get publication by price
      description: Get publications with the price indicated
      operationId: getPublicationByPrice
      parameters:
        - $ref: '#/components/parameters/price'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /publications/price/between:
    get:
      tags:
        - Publication
      summary: Get publication by price between
      description: Get publications between the two prices indicated.
      operationId: getPublicationByPriceBetween
      parameters:
        - $ref: '#/components/parameters/minPrice'
        - $ref: '#/components/parameters/maxPrice'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /publications/score/{score}:
    get:
      tags:
        - Publication
      summary: Get publication by score
      description: Get publications with the score indicated. The score goes from 0 to 5.
      operationId: getPublicationsByScore
      parameters:
        - $ref: '#/components/parameters/score'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /publications/score/between:
    get:
      tags:
        - Publication
      summary: Get publication by score between
      description: Get publications between the two scores indicated. The score goes from 0 to 5.
      operationId: getPublicationByScoreBetween
      parameters:
        - $ref: '#/components/parameters/minScore'
        - $ref: '#/components/parameters/maxScore'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /publications/establishment/{id}:
    get:
      tags:
        - Publication
      summary: Get publication by establishment id
      description: Get publications that belongs to the establishment indicated
      operationId: getPublicationByEstablishmentId
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /publications/user/{id}:
    get:
      tags:
        - Publication
      summary: Get publication by user id
      description: Get publications that belongs to the user indicated.
      operationId: getPublicationByUserId
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        404:
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
  /publications/type/{type}:
    get:
      tags:
        - Publication
      summary: Get publication by product type
      description: Get publications that belongs to the product type indicated.
      operationId: getPublicationByProductType
      parameters:
        - $ref: '#/components/parameters/type'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publication'
        401:
          description: Unauthorize exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
...
components:
  schemas:
...
    Publication:
      properties:
        id:
          type: integer
          format: int64
          example: 0
        date:
          type: string
          format: dd-MM-yyyy
          example: 03-05-2022
        totalPrice:
          type: number
          example: 3.4
        totalScore:
          type: number
          example: 2.5
        photo:
          type: string
        user:
          $ref: '#/components/schemas/User'
        establishment:
          $ref: '#/components/schemas/Establishment'            
    PublicationDTO:
      properties:
        photo:
          type: string
          example: tostada.png
        userId:
          type: integer
          format: int64
          example: 5
        establishmentId:
          type: integer
          format: int64   
          example: 3
...
    Exception:
      properties:
        code: 
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
        message:
          type: string
  parameters:
    id:
      name: id
      in: path
      description: id of the element
      required: true
      schema:
        type: integer
        format: int64
        example: 0
    date:
      name: date
      in: path
      description: creation date
      required: true
      schema:
        type: integer
        example: 1640604419000
    minDate:
      name: minDate
      in: query
      description: date from in timestamp miliseconds
      required: true
      schema:
        type: integer
        example: 1640604419000
    maxDate:
      name: maxDate
      in: query
      description: date to in timestamp miliseconds
      required: true
      schema:
        type: integer
        example: 1641604419000
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT